Allow to pass URL instead of the name of the Pod
A somewhat naive implementation of adding URL handling so you can pass
in an http/https/github-ssh based URL to pod try.

I didn’t want to stomp on the cocoapods-try-release-fix-gemspec because
it was just added back to the repo, but I did specify an explicit name
in the Gemfile.