Move from podspec_location to podspec_name
The flag is poorly named, this is more clear and the other
functionality is not supported of using a subdirectory. Only
a .podspec file of a different name than the git repo is
supported, and this corrects the functionality based on that.

The podspec file still has to be located in the root of the git repo.